update postman collection + email service config

+++ backend/src/main/java/shelter/backend/registration/service/ShelterRegistrationService.java
@@ -5,9 +5,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import shelter.backend.admin.model.AdminResponseConverter;
-import shelter.backend.admin.model.AdminShelterRequestDto;
-import shelter.backend.admin.model.AdminShelterResponseDto;
 import shelter.backend.email.EmailService;
 import shelter.backend.rest.model.mapper.UserMapper;
 import shelter.backend.rest.model.dtos.UserDto;
@@ -20,10 +17,12 @@ import shelter.backend.storage.repository.AddressRepository;
 import shelter.backend.storage.repository.RoleRepository;
 import shelter.backend.storage.repository.UserRepository;
 import shelter.backend.rest.model.entity.User;
+import shelter.backend.utils.exception.MessageNotSendException;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
+import java.util.Optional;
 import java.util.stream.StreamSupport;
 
 @Service
@@ -54,19 +53,8 @@ public class ShelterRegistrationService implements RegistrationService {
         this.userMapper = new UserMapper(userRepository, addressRepository);
     }
 
-    @Override
-    public UserDto registerUser(UserDto userDto) {
-        log.debug("User Registration started for username: {}", userDto.getEmail());
-        return register(userDto);
-    }
-
-    @Override
-    public UserDto registerShelter(UserDto userDto) {
-        log.debug("Shelter Registration started for username: {}", userDto.getEmail());
-        return register(userDto);
-    }
-
     public UserDto register(UserDto userDto) {
+        log.debug("Registration started for username: {}", userDto.getEmail());
         userValidator.throwIfNotValid(userDto);
         User user = persistTheUser(userDto);
         log.info("1st stage of registration for user: {} completed successfully", user.getEmail());
@@ -77,12 +65,18 @@ public class ShelterRegistrationService implements RegistrationService {
     private void sendConfirmationEmail(User user) {
         Token token = tokenService.generateToken(user.hashCode(), user.getEmail());
         log.debug("Confirmation token {} created for user: {}", token.getId(), user.getEmail());
+        try {
+
         if (isShelter(user)) {
             shelterEmailService.sendShelterConfirmationEmail(user.getEmail(), token.getId());
-            user.setApprovalStatus(ApprovalStatus.EMAIL_NIEPOTWIERDZONY);//TODO na angielski
+            user.setApprovalStatus(ApprovalStatus.EMAIL_NOT_CONFIRMED);
             userRepository.save(user);
         } else {
             shelterEmailService.sendUserConfirmationEmail(user.getEmail(), token.getId());
+        }
+        }catch (MessageNotSendException e){
+            userRepository.deleteById(user.getId());
+            throw e;
         }
     }
 
@@ -104,7 +98,7 @@ public class ShelterRegistrationService implements RegistrationService {
         User user = userRepository.findUserByEmail(token.getUsername());
 
         if (isShelter(user)) {
-            user.setApprovalStatus(ApprovalStatus.OCZEKUJE);
+            user.setApprovalStatus(ApprovalStatus.PENDING);
             checkApprovalStatus(user);
         } else {
             user.setDisabled(false);
@@ -120,9 +114,9 @@ public class ShelterRegistrationService implements RegistrationService {
         try {
             boolean isApproved = approvalProvider.validateShelterDetails(user.getAddress());
             if (isApproved) {
-                user.setApprovalStatus(ApprovalStatus.POTWIERDZONY);
+                user.setApprovalStatus(ApprovalStatus.CONFIRMED);
             } else {
-                user.setApprovalStatus(ApprovalStatus.ODRZUCONY);
+                user.setApprovalStatus(ApprovalStatus.REJECTED);
             }
         } catch (Exception e) {
             log.error("Excepetion occured during validating shelter details. ", e);
@@ -130,37 +124,46 @@ public class ShelterRegistrationService implements RegistrationService {
     }
 
     @Override
-    public List<AdminShelterResponseDto> enableShelterAccounts(List<AdminShelterRequestDto> shelterList) { //TODO zmiana response na liste enablowanych shelterów
-        shelterList.forEach(shelter -> {
-            User user = userRepository.findUserByEmail(shelter.getEmail());
-            if (user != null) {
-                log.info("Shelter: {}, for username: {} accepted by admin", shelter.getShelterName(), shelter.getEmail());
-                user.setDisabled(false);
-                userRepository.save(user);
+    public List<UserDto> enableShelterAccounts(List<Long> shelterIdList) {
+        List<UserDto> enabledShelters = new ArrayList<>();
+        shelterIdList.forEach(id -> {
+            Optional<User> userOptional = userRepository.findById(id);
+            if (userOptional.isPresent()) {
+                User user = userOptional.get();
+                if (user.isDisabled()) {
+                    log.info("Shelter: {}, for username: {} accepted by admin", user.getShelterName(), user.getEmail());
+                    user.setDisabled(false);
+                    userRepository.save(user);
+                    enabledShelters.add(userMapper.toDto(user));
+                }
+                log.debug("Shelter for username {} already enabled", user.getEmail());
             }
         });
-        List<User> disabledShelters = userRepository.findAllByShelterNameIsNotNullAndIsDisabledIsTrue();
-        return disabledShelters.stream()
-                .map(AdminResponseConverter::toAdminGetShelterResponse)
-                .collect(Collectors.toList());
+        return enabledShelters;
     }
 
     @Transactional
     private void removeExpTokenAndAssocUser(Token token, String username) {
         User user = userRepository.findUserByEmail(username);
-        if (user.isDisabled()) {
+        if (user != null && user.isDisabled()) {
             userRepository.deleteUserByEmail(username);
-            log.debug("Deleted username: {}", username);
+            log.debug("[SCHEDULER] :: Username: {} deleted", username);
         }
         tokenService.deleteToken(token);
-        log.debug("Deletet token: {}", token);
+        log.debug("[SCHEDULER] :: Token: {} deleted", token);
     }
 
     private User persistTheUser(UserDto userDto) {
         User newUser = userMapper.toEntity(userDto);
         newUser.setPassword(passwordEncoder.encode(userDto.getPassword()));
         newUser.setDisabled(true);
-        Role role = roleRepository.findByName(ERole.ROLE_USER);
+        ERole roleName;
+        if (isShelter(newUser)) {
+            roleName = ERole.ROLE_SHELTER;
+        } else {
+            roleName = ERole.ROLE_USER;
+        }
+        Role role = roleRepository.findByName(roleName);
         newUser.getRoles().add(role);
         return userRepository.save(newUser);
     }
